requirements-test.txt
-r requirements.txt
pytest==3.4.2
mock==2.0.0
case==1.5.3
mccabe==0.6.1
pep8-naming==0.5.0
flake8==3.5.0
coverage>=4.0,<4.4
pytest-cov==2.5.1
tox==2.9.1
pyramid
pyramid_celery
